requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
websocket-client>=1.7.0
tzdata>=2024.1
